:recycle: refactoring api of front
